+++ test/app/components/slim_with_unsafe_html_component.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class SlimWithUnsafeHTMLComponent < ViewComponent::Base
+class SlimWithUnsafeHtmlComponent < ViewComponent::Base
   def unsafe_html
     "<script>alert('xss')</script>"
   end